+++ iqm_client/src/iqm/qiskit_iqm/fake_backends/iqm_fake_backend.py
@@ -1,4 +1,4 @@
-# Copyright 2022-2023 Qiskit on IQM developers
+# Copyright 2022-2025 Qiskit on IQM developers
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -18,9 +18,16 @@ from __future__ import annotations
 from copy import deepcopy
 from dataclasses import dataclass
 from itertools import permutations
-from typing import Any, Optional, Union
+from typing import Optional, Union
+from uuid import UUID
 
-from iqm.iqm_client import DynamicQuantumArchitecture, QuantumArchitectureSpecification
+from iqm.iqm_client import (
+    DynamicQuantumArchitecture,
+    GateImplementationInfo,
+    GateInfo,
+    Locus,
+    StaticQuantumArchitecture,
+)
 from iqm.qiskit_iqm.iqm_backend import IQM_TO_QISKIT_GATE_NAME, IQMBackendBase
 from iqm.qiskit_iqm.iqm_circuit_validation import validate_circuit
 from iqm.qiskit_iqm.iqm_transpilation import IQMReplaceGateWithUnitaryPass
@@ -28,20 +35,70 @@ from iqm.qiskit_iqm.move_gate import MOVE_GATE_UNITARY
 from qiskit import QuantumCircuit
 from qiskit.providers import JobV1, Options
 from qiskit_aer import AerSimulator
-from qiskit_aer.noise import NoiseModel
+from qiskit_aer.noise import NoiseModel, QuantumError
 from qiskit_aer.noise.errors import depolarizing_error, thermal_relaxation_error
+
+
+def _dqa_from_sqa(
+    sqa: StaticQuantumArchitecture,
+    error_profile: IQMErrorProfile,
+) -> DynamicQuantumArchitecture:
+    """Create a dynamic quantum architecture from the given static quantum architecture
+    and error profile.
+
+    Since the DQA contains some attributes that are not present in an SQA or error profile,
+    they are filled with mock data:
+
+    * Each gate type is given a single mock implementation.
+    * Calibration set ID is set to the all-zeros UUID.
+
+    Args:
+        sqa: Static quantum architecture to replicate.
+        error_profile: Characteristics of a particular QPU specimen (including gate loci).
+
+    Returns:
+        DQA replicating the properties of ``sqa``.
+
+    """
+
+    def fake_gateinfo(loci: tuple[Locus, ...]) -> GateInfo:
+        """Fake GateInfo for the given loci."""
+        return GateInfo(
+            implementations={"__fake": GateImplementationInfo(loci=loci)},
+            default_implementation="__fake",
+            override_default_implementation={},
+        )
+
+    # add all gates/loci found in the error profile
+    gates = {
+        gate_name: fake_gateinfo(tuple((component,) for component in gate_errors_1q))
+        for gate_name, gate_errors_1q in error_profile.single_qubit_gate_depolarizing_error_parameters.items()
+    }
+
+    for gate_name, gate_errors_2q in error_profile.two_qubit_gate_depolarizing_error_parameters.items():
+        gates[gate_name] = fake_gateinfo(tuple(gate_errors_2q))
+
+    gates["measure"] = fake_gateinfo(tuple((component,) for component in error_profile.readout_errors))
+
+    return DynamicQuantumArchitecture(
+        calibration_set_id=UUID("00000000-0000-0000-0000-000000000000"),
+        qubits=sqa.qubits,
+        computational_resonators=sqa.computational_resonators,
+        gates=gates,
+    )
 
 
 # pylint: disable=too-many-instance-attributes
 @dataclass
 class IQMErrorProfile:
-    r"""Properties of an IQM QPU specimen, used for constructing an error model.
+    r"""Characteristics of an IQM QPU specimen, used for constructing an error model.
 
-    All the attributes of this class refer to the qubits of the QPU using their physical names.
+    All the attributes of this class refer to the components of the QPU using their physical names.
+    There are two types of QPU components, qubits and computational resonators.
 
     Args:
-        t1s: maps qubits to their :math:`T_1` times (in ns)
-        t2s: maps qubits to the :math:`T_2` times (in ns)
+        t1s: maps components to their :math:`T_1` times (in ns)
+        t2s: maps components to their :math:`T_2` times (in ns)
         single_qubit_gate_depolarizing_error_parameters: Depolarizing error parameters for single-qubit gates.
             Maps single-qubit gate names to a mapping of qubits (on which the gate acts) to a depolarizing error.
             The error, used in a one-qubit depolarizing channel, concatenated with a thermal relaxation channel,
@@ -63,9 +120,9 @@ class IQMErrorProfile:
             IQMErrorProfile(
                 t1s={"QB1": 10000.0, "QB2": 12000.0, "QB3": 14000.0},
                 t2s={"QB1": 10000.0, "QB2": 12000.0, "QB3": 13000.0},
-                single_qubit_gate_depolarizing_error_parameters={"r": {"QB1": 0.0005, "QB2": 0.0004, "QB3": 0.0010}},
+                single_qubit_gate_depolarizing_error_parameters={"prx": {"QB1": 0.0005, "QB2": 0.0004, "QB3": 0.0010}},
                 two_qubit_gate_depolarizing_error_parameters={"cz": {("QB1", "QB2"): 0.08, ("QB2", "QB3"): 0.03}},
-                single_qubit_gate_durations={"r": 50.},
+                single_qubit_gate_durations={"prx": 50.},
                 two_qubit_gate_durations={"cz": 100.},
                 readout_errors={"QB1": {"0": 0.02, "1": 0.03},
                                 "QB2": {"0": 0.02, "1": 0.03},
@@ -84,41 +141,46 @@ class IQMErrorProfile:
     readout_errors: dict[str, dict[str, float]]
     name: Optional[str] = None
 
+    def thermal_relaxation(self, component: str, duration: float) -> QuantumError:
+        """One-qubit relaxation error channel."""
+        return thermal_relaxation_error(self.t1s[component], self.t2s[component], duration)
+
 
 class IQMFakeBackend(IQMBackendBase):
     """Simulated backend that mimics the behaviour of IQM quantum computers.
 
     Can be used to perform noisy gate-level simulations of quantum circuit execution on IQM hardware.
 
-    A fake backend contains information about a specific IQM system, such as the quantum architecture (number of qubits,
-    connectivity), the native gate set, and a noise model based on system parameters such as relaxation (:math:`T_1`)
+    A fake backend contains information about a specific IQM system, such as the static quantum architecture
+    (number of qubits, connectivity), and a noise model based on system parameters such as relaxation (:math:`T_1`)
     and dephasing (:math:`T_2`) times, gate infidelities, and readout errors.
 
     Args:
-        architecture: Quantum architecture associated with the backend instance.
+        architecture: Static quantum architecture associated with the backend instance.
         error_profile: Characteristics of a particular QPU specimen.
 
     """
 
     def __init__(
         self,
-        architecture: QuantumArchitectureSpecification,
+        architecture: StaticQuantumArchitecture,
         error_profile: IQMErrorProfile,
         name: str = "IQMFakeBackend",
         **kwargs,
     ):
-        super().__init__(architecture, **kwargs)
-
         self._validate_architecture_and_error_profile(architecture, error_profile)
-        self.__architecture, self.__error_profile = architecture, error_profile
+        dqa = _dqa_from_sqa(architecture, error_profile)
+        super().__init__(dqa, **kwargs)
+        self.__sqa = architecture
+        self.__dqa = dqa
+        self.__error_profile = error_profile
 
-        self.noise_model = self._create_noise_model(architecture, error_profile)
-
+        self.noise_model = self._create_noise_model()
         self.name = name
 
     @property
     def error_profile(self) -> IQMErrorProfile:
-        """Error profile of this instance of IQM fake backend"""
+        """Error profile of this IQM fake backend instance."""
         return deepcopy(self.__error_profile)
 
     @error_profile.setter
@@ -132,66 +194,54 @@ class IQMFakeBackend(IQMBackendBase):
         """Return another instance of IQMFakeBackend, which has the same quantum architecture but a different error
         profile.
         """
-        return self.__class__(self.__architecture, new_error_profile)
+        return self.__class__(self.__sqa, new_error_profile, self.name)
 
     @staticmethod
     def _validate_architecture_and_error_profile(
-        architecture: QuantumArchitectureSpecification, error_profile: IQMErrorProfile
+        architecture: StaticQuantumArchitecture, error_profile: IQMErrorProfile
     ) -> None:
-        """Verifies that the parameters of the QPU error profile match the constraints of its quantum architecture.
+        """Verify that the parameters of the QPU error profile match the constraints of its quantum architecture.
 
         Raises:
-            ValueError: when length of `t1s` and number of qubits do not match.
-            ValueError: when length of `t2s` and number of qubits do not match.
-            ValueError: when length of `one_qubit_gate` parameter lists and number of qubits do not match.
-            ValueError: when length of `two_qubit_gate` parameter lists and number of couplings do not match.
+            ValueError: when length of :attr:`t1s` and number of qubits do not match.
+            ValueError: when length of :attr:`t2s` and number of qubits do not match.
+            ValueError: when length of :attr:`one_qubit_gate` parameter lists and number of qubits do not match.
+            ValueError: when length of :attr:`two_qubit_gate` parameter lists and number of couplings do not match.
             ValueError: when gates in gate parameter lists are not supported by the quantum architecture.
 
         """
-        num_qubits = len(architecture.qubits)
-        # Check that T1 list has one element for each qubit
-        if len(error_profile.t1s) != num_qubits:
-            raise ValueError(
-                f"Length of t1s ({len(error_profile.t1s)}) and number of qubits ({num_qubits}) should match."
-            )
 
-        # Check that T2 list has one element for each qubit
-        if len(error_profile.t2s) != num_qubits:
-            raise ValueError(
-                f"Length of t2s ({len(error_profile.t2s)}) and number of qubits ({num_qubits}) should match."
-            )
+        def compare_sets(a: set, b: set, msg: str) -> None:
+            """Compare the contents of two sets."""
+            if diff := a - b:
+                raise ValueError(f"Error profile attribute {msg} has unknown loci: {diff}.")
+            if diff := b - a:
+                raise ValueError(f"Error profile attribute {msg} is missing loci: {diff}.")
 
-        property_dict: dict[str, dict[Any, float]]
+        components = set(architecture.qubits + architecture.computational_resonators)
+        # Check that T1 and T2 dicts have one item for each component
+        compare_sets(set(error_profile.t1s.keys()), components, "t1s")
+        compare_sets(set(error_profile.t2s.keys()), components, "t2s")
+
         # Check that one-qubit gate parameter qubits match those of the architecture
-        for property_name, property_dict in [
-            ("depolarizing rates", error_profile.single_qubit_gate_depolarizing_error_parameters),
-        ]:
-            gate_dict: dict[Any, float]
-            for gate, gate_dict in property_dict.items():
-                if set(gate_dict.keys()) != set(architecture.qubits):
-                    raise ValueError(
-                        (
-                            f"The qubits specified for one-qubit gate {property_name} ({set(gate_dict.keys())}) "
-                            f"don't match the qubits of the quantum architecture "
-                            f"`{architecture.name}` ({architecture.qubits})."
-                        )
-                    )
+        for gate, gate_errors_1q in error_profile.single_qubit_gate_depolarizing_error_parameters.items():
+            compare_sets(
+                set(gate_errors_1q.keys()),
+                set(architecture.qubits),
+                f"single_qubit_gate_depolarizing_error_parameters['{gate}']",
+            )
 
         # Check that two-qubit gate parameter couplings match those of the architecture
-        for property_name, property_dict in [
-            ("depolarizing error parameters", error_profile.two_qubit_gate_depolarizing_error_parameters),
-        ]:
-            for gate, gate_dict in property_dict.items():
-                if set(gate_dict.keys()) != set(tuple(item) for item in architecture.qubit_connectivity):
-                    raise ValueError(
-                        (
-                            f"The couplings specified for two-qubit gate {property_name} ({set(gate_dict.keys())}) "
-                            f"don't match the couplings of the quantum architecture "
-                            f"`{architecture.name}` ({architecture.qubit_connectivity})."
-                        )
-                    )
+        # Locus order is ignored because the loci in architecture.connectivity are unordered.
+        for gate, gate_errors_2q in error_profile.two_qubit_gate_depolarizing_error_parameters.items():
+            compare_sets(
+                set(frozenset(locus) for locus in gate_errors_2q.keys()),
+                set(frozenset(locus) for locus in architecture.connectivity),
+                f"two_qubit_gate_depolarizing_error_parameters['{gate}']",
+            )
 
-        # Check that the basis gates of the chip sample match the quantum architecture's
+        # Check that the basis gates in the error profile are all known
+        known_gates = {"prx", "cc_prx", "cz", "move"}
         for property_name, specified_gates in [
             (
                 "single_qubit_gate_depolarizing_error_parameters",
@@ -204,63 +254,46 @@ class IQMFakeBackend(IQMBackendBase):
             ("durations", (error_profile.single_qubit_gate_durations | error_profile.two_qubit_gate_durations).keys()),
         ]:
             for gate in specified_gates:
-                if gate not in architecture.operations:
-                    raise ValueError(
-                        (
-                            f"Gate `{gate}` in `{property_name}` "
-                            f"is not supported by quantum architecture `{architecture.name}`. "
-                            f"Valid gates: {architecture.operations}"
-                        )
-                    )
-        if set(error_profile.readout_errors.keys()) != set(architecture.qubits):
-            raise ValueError(
-                f"The qubits specified in readout errors ({set(error_profile.readout_errors.keys())}) "
-                f"don't match the qubits of the quantum architecture "
-                f"`{architecture.name}` ({architecture.qubits})."
-            )
+                if gate not in known_gates:
+                    raise ValueError(f"Gate '{gate}' in {property_name} is unknown. Valid gates: {known_gates}")
 
-    def _create_noise_model(
-        self, architecture: QuantumArchitectureSpecification, error_profile: IQMErrorProfile
-    ) -> NoiseModel:
-        """Builds a noise model from the attributes."""
-        iqm_to_qiskit_gates = dict(IQM_TO_QISKIT_GATE_NAME)
-        for iqm_gate in architecture.operations:
-            if iqm_gate not in ["measure", "barrier"] + list(iqm_to_qiskit_gates):
-                iqm_to_qiskit_gates[iqm_gate] = iqm_gate
+        compare_sets(set(error_profile.readout_errors.keys()), set(architecture.qubits), "readout_errors")
+
+    def _create_noise_model(self) -> NoiseModel:
+        """Build a noise model from the attributes."""
+        error_profile = self.__error_profile
+        iqm_to_qiskit_gates = IQM_TO_QISKIT_GATE_NAME.copy()
+        for iqm_gate in self.__dqa.gates:
+            if iqm_gate not in ["measure", "barrier"]:
+                iqm_to_qiskit_gates.setdefault(iqm_gate, iqm_gate)
+
         noise_model = NoiseModel(basis_gates=list(iqm_to_qiskit_gates.values()))
+
         # Add single-qubit gate errors to noise model
-        for gate in error_profile.single_qubit_gate_depolarizing_error_parameters.keys():
-            for qb in architecture.qubits:
-                thermal_relaxation_channel = thermal_relaxation_error(
-                    error_profile.t1s[qb], error_profile.t2s[qb], error_profile.single_qubit_gate_durations[gate]
-                )
-                depolarizing_channel = depolarizing_error(
-                    error_profile.single_qubit_gate_depolarizing_error_parameters[gate][qb], 1
-                )
-                full_error_channel = thermal_relaxation_channel.compose(depolarizing_channel)
+        for gate, gate_errors_1q in error_profile.single_qubit_gate_depolarizing_error_parameters.items():
+            gate_duration = error_profile.single_qubit_gate_durations[gate]
+            for component, gate_error in gate_errors_1q.items():
+                thermal_relaxation_channel = error_profile.thermal_relaxation(component, gate_duration)
+                depolarizing_channel = depolarizing_error(gate_error, 1)
                 noise_model.add_quantum_error(
-                    full_error_channel, iqm_to_qiskit_gates[gate], [self.qubit_name_to_index(qb)]
+                    thermal_relaxation_channel.compose(depolarizing_channel),
+                    iqm_to_qiskit_gates[gate],
+                    [self.qubit_name_to_index(component)],
                 )
 
         # Add two-qubit gate errors to noise model
-        for gate, rates in error_profile.two_qubit_gate_depolarizing_error_parameters.items():
-            for (qb1, qb2), rate in rates.items():
-                for qb_order in permutations((qb1, qb2)):
-                    thermal_relaxation_channel = thermal_relaxation_error(
-                        error_profile.t1s[qb_order[0]],
-                        error_profile.t2s[qb_order[0]],
-                        error_profile.two_qubit_gate_durations[gate],
-                    ).tensor(
-                        thermal_relaxation_error(
-                            error_profile.t1s[qb_order[1]],
-                            error_profile.t2s[qb_order[1]],
-                            error_profile.two_qubit_gate_durations[gate],
-                        )
-                    )
-                    depolarizing_channel = depolarizing_error(rate, 2)
-                    full_error_channel = thermal_relaxation_channel.compose(depolarizing_channel)
+        for gate, gate_errors_2q in error_profile.two_qubit_gate_depolarizing_error_parameters.items():
+            gate_duration = error_profile.two_qubit_gate_durations[gate]
+            for locus, gate_error in gate_errors_2q.items():
+                for qb_order in permutations(locus):
+                    # TODO why do we need to add the other locus order?
+                    thermal_channels = [
+                        error_profile.thermal_relaxation(component, gate_duration) for component in qb_order
+                    ]
+                    thermal_relaxation_channel = thermal_channels[0].tensor(thermal_channels[1])
+                    depolarizing_channel = depolarizing_error(gate_error, 2)
                     noise_model.add_quantum_error(
-                        full_error_channel,
+                        thermal_relaxation_channel.compose(depolarizing_channel),
                         iqm_to_qiskit_gates[gate],
                         [self.qubit_name_to_index(qb_order[0]), self.qubit_name_to_index(qb_order[1])],
                     )
@@ -324,27 +357,15 @@ class IQMFakeBackend(IQMBackendBase):
         job = sim_noise.run(circuits, shots=shots)
         return job
 
-    def validate_compatible_architecture(self, architecture: DynamicQuantumArchitecture) -> bool:
-        """Compare a dynamic quantum architecture to the static architecture of the fake backend.
+    def validate_compatible_architecture(self, architecture: StaticQuantumArchitecture) -> bool:
+        """Compare a static quantum architecture to the static architecture of the fake backend.
 
         Args:
-            architecture: dynamic quantum architecture to compare to
+            architecture: static quantum architecture to compare to
 
         Returns:
-            True iff the number and names of the locus components, the component connectivity,
-            and the available operations in the DQA match the static architecture of this backend.
+            True iff the locus components and the component connectivity
+            in the IQMFakeBackend SQA match ``architecture``.
 
         """
-        components_match = set(architecture.components) == set(self.__architecture.qubits)
-        ops = {
-            gate_name: list(list(locus) for locus in gate_info.loci)
-            for gate_name, gate_info in architecture.gates.items()
-        }
-        ops.update({"barrier": []})
-        ops_match = self.__architecture.compare_operations(self.__architecture.operations, ops)
-
-        self_connectivity = set(map(frozenset, self.__architecture.qubit_connectivity))
-        target_connectivity = set(frozenset(locus) for loci in ops.values() for locus in loci if len(locus) > 1)
-        connectivity_match = self_connectivity == target_connectivity
-
-        return components_match and ops_match and connectivity_match
+        return architecture == self.__sqa