+++ iqm_client/src/iqm/qiskit_iqm/iqm_circuit_validation.py
@@ -33,7 +33,7 @@ def validate_circuit(
         qubit_mapping = backend._idx_to_qb
     new_circuit = IQMClientCircuit(
         name="Validation circuit",
-        instructions=serialize_instructions(circuit=circuit, qubit_index_to_name=qubit_mapping),
+        instructions=tuple(serialize_instructions(circuit=circuit, qubit_index_to_name=qubit_mapping)),
     )
     if validate_moves is None:
         validate_moves = MoveGateValidationMode.STRICT